# docs/openapi.yaml
openapi: 3.0.1
info:
  title: Team-CBT API
  description: "This is an API for Team-CBT web app"
  version: 1.0.0
servers:
  - url: https://teamcbt.com/v1
  - url: http://teamcbt.com/v1
...
paths:
...
  /user:
    post:
      tags:
        - user
      summary: Register user
      description: Register user
      operationId: createUser
      requestBody:
        description: Created user object
        content:
          "*/*":
            schema:
              $ref: "#/components/schemas/RegisterUser"
        required: true
      responses:
        200:
          description: successful operation
          content: {}
        409:
          description: Operation failed or User already exists
          content: {}
        503:
          description: Server is currently down
          content: {}
      x-codegen-request-body-name: body
    get:
      tags:
        - user
      summary: get list of users
      operationId: getAllUsers
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        403:
          description: Access Forbidden
          content: {}
        500:
          description: Failure of users retrieval
          content: {}
        503:
          description: Server is currently down
          content: {}
      security:
        - teamcbt_auth:
            - write:session
            - read:session
...
  /user/{userId}:
    get:
      tags:
        - user
      summary: Get user by userId
      operationId: getUserByUserId
      parameters:
        - name: userId
          in: path
          description: "The ID of the user that needs to be fetched. Use 1 for testing. "
          required: true
          schema:
            type: integer
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        400:
          description: Invalid email supplied
          content: {}
        403:
          description: Access Forbidden
          content: {}
        404:
          description: User not found
          content: {}
        503:
          description: Server is currently down
          content: {}
    put:
      tags:
        - user
      summary: Update user
      description: This can only be done by the logged in user.
      operationId: updateUser
      parameters:
        - name: userId
          in: path
          description: name that need to be updated
          required: true
          schema:
            type: integer
      requestBody:
        description: Updated user object
        content:
          "*/*":
            schema:
              $ref: "#/components/schemas/RegisterUser"
        required: true
      responses:
        200:
          description: User updated.
          content:
            application/json:
              schema:
                properties:
                  login:
                    type: string
                  userStatus:
                    type: integer
                  accessToken:
                    type: string

        400:
          description: Invalid user supplied
          content: {}
        403:
          description: Access Forbidden
          content: {}
        404:
          description: User not found
          content: {}
        500:
          description: failure
          content: {}
        503:
          description: Server is currently down
          content: {}
      x-codegen-request-body-name: body
    delete:
      tags:
        - user
      summary: Delete user
      description: This can only be done by the logged in user.
      operationId: deleteUser
      parameters:
        - name: userId
          in: path
          description: The ID that needs to be deleted
          required: true
          schema:
            type: integer
      responses:
        204:
          description: User deleted
          content: {}
        400:
          description: Invalid username supplied
          content: {}
        403:
          description: Access Forbidden
          content: {}
        404:
          description: User not found
          content: {}
        500:
          description: Failure
          content: {}
        503:
          description: Server is currently down
          content: {}
...
components:
  schemas:
    RegisterUser:
      required:
        - login
        - password
        - age
        - gender
      properties:
        login:
          type: string
        password:
          type: string
        age:
          type: integer
        gender:
          type: string
...
    User:
      required:
        - userId
        - login
        - userStatus
        - banned
      type: object
      properties:
        userId:
          type: integer
        login:
          type: string
        age:
          type: integer
        gender:
          type: string
        userStatus:
          type: integer
        banned:
          type: boolean
...
  securitySchemes:
    teamcbt_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    api_key:
      type: apiKey
      name: api_key
      in: header